% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{add_absolute_proportions_to_tree_from_leaves}
\alias{add_absolute_proportions_to_tree_from_leaves}
\title{Calculate absolute proportions from relative proportions}
\usage{
add_absolute_proportions_to_tree_from_leaves(node)
}
\description{
Calculate absolute proportions from relative proportions
}
