% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tree.R
\name{add_absolute_proportions_to_trees}
\alias{add_absolute_proportions_to_trees}
\title{Put absolute proportions on trees}
\usage{
add_absolute_proportions_to_trees(trees)
}
\description{
Put absolute proportions on trees
}
