% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_hierarchical.R
\name{add_cox_test}
\alias{add_cox_test}
\title{add_cox_test}
\usage{
add_cox_test(.data, relative = TRUE)
}
\arguments{
\item{.data}{A tibble}

\item{relative}{A boolean}
}
\description{
This function adds cox regression statistics to the fit object
}
