% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{add_data_to_tree_from_table}
\alias{add_data_to_tree_from_table}
\title{Put proportions on tree}
\usage{
add_data_to_tree_from_table(tree, proportion, label, append)
}
\description{
Put proportions on tree
}
