% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ancestor_child}
\alias{ancestor_child}
\title{Data frame of the link between nodes of the cell tree}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 2 columns.
}
\usage{
ancestor_child
}
\description{
Data frame of the link between nodes of the cell tree
}
\keyword{datasets}
