% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ARMET_ref}
\alias{ARMET_ref}
\title{Reference signatures}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9036739 rows and 14 columns.
}
\usage{
ARMET_ref
}
\description{
Reference signatures
}
\keyword{datasets}
