% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_tc.R
\name{ARMET_tc_continue}
\alias{ARMET_tc_continue}
\title{ARMET_tc_continue}
\usage{
ARMET_tc_continue(
  armet_obj,
  level,
  model = stanmodels$ARMET_tc_fix_hierarchical
)
}
\arguments{
\item{armet_obj}{An ARMEt object}

\item{level}{An integer}

\item{model}{A stan model}
}
\description{
This function does inference for higher levels of the hierarchy
}
