% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{array_norm}
\alias{array_norm}
\title{Normalize array data of ref and mix}
\usage{
array_norm(ref, mix)
}
\arguments{
\item{ref}{A matrix}

\item{mix}{A matrix}
}
\value{
A list including the ref and mix normalized and a ggplot
}
\description{
Normalize array data of ref and mix
}
