% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{average_duplicated_genes}
\alias{average_duplicated_genes}
\title{Remove gene redundancy averagin duplicated genes}
\usage{
average_duplicated_genes(ref, symbol)
}
\arguments{
\item{ref}{A matrix}

\item{symbol}{A char vector}
}
\value{
Non redundant matrix
}
\description{
Remove gene redundancy averagin duplicated genes
}
