% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{check_if_sd_zero_and_correct}
\alias{check_if_sd_zero_and_correct}
\title{Checks if the standard deviation of a gene is 0}
\usage{
check_if_sd_zero_and_correct(df, node)
}
\arguments{
\item{df}{A matrix}
}
\description{
Checks if the standard deviation of a gene is 0
}
