% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{check_input}
\alias{check_input}
\title{Check the input for anomalies}
\usage{
check_input(mix, is_mix_microarray, my_design, cov_to_test, custom_ref, tree)
}
\description{
Check the input for anomalies
}
