% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{check_markers}
\alias{check_markers}
\title{Check if prediction of marker value in full ayesian model is too different from the provided reference marker values distibutions}
\usage{
check_markers(fit, map)
}
\arguments{
\item{fit}{A stan fit object}

\item{map}{A matrix}
}
\description{
Check if prediction of marker value in full ayesian model is too different from the provided reference marker values distibutions
}
