% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{choose_chains_majority_roule}
\alias{choose_chains_majority_roule}
\title{choose_chains_majority_roule}
\usage{
choose_chains_majority_roule(fit_parsed)
}
\description{
Get which chain cluster is more opulated in case I have divergence
}
