% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmdstanr_ARMET_tc_fix_hierarchical}
\alias{cmdstanr_ARMET_tc_fix_hierarchical}
\title{Model for cmdstanr}
\format{
An object of class \code{character} of length 1.
}
\usage{
cmdstanr_ARMET_tc_fix_hierarchical
}
\description{
Model for cmdstanr
}
\keyword{datasets}
