% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{convoluted_glm}
\alias{convoluted_glm}
\title{convoluted_glm main}
\usage{
convoluted_glm(
  .data,
  .formula = ~1,
  .sample,
  .transcript,
  .abundance,
  reference = NULL,
  tree = NULL,
  approximate_posterior = F,
  prior_survival_time = c(),
  transform_time_function = sqrt,
  use_data = TRUE,
  use_cmdstanr = FALSE
)
}
\arguments{
\item{.data}{A tibble including a cell_group name column | sample name column | read counts column (optional depending on the input class) | covariate columns.}

\item{.sample}{A column name as symbol. The sample identifier}

\item{.transcript}{A column name as symbol. The cell_group identifier}

\item{.abundance}{A column name as symbol. The cell_group abundance (read count). Used only for data frame count output. The variable in this column should be of class integer.}

\item{reference}{A data frame}

\item{tree}{A node object}

\item{approximate_posterior}{A boolean}

\item{prior_survival_time}{A list}

\item{transform_time_function}{A function with nake survival time normally-shaped.}

\item{formula}{A formula. The formula describing the model for differential abundance, for example ~treatment.}
}
\value{
A nested tibble `tbl`, with the following columns
\itemize{
  \item cell_group - column including the cell groups being tested
  \item parameter - The parameter being estimated, from the design matrix dscribed with the input formula_composition and formula_variability

  \item c_lower - lower (2.5%) quantile of the posterior distribution for a composition (c) parameter.
  \item c_effect - mean of the posterior distribution for a composition (c) parameter.
  \item c_upper - upper (97.5%) quantile of the posterior distribution fo a composition (c)  parameter.
  \item c_pH0 - Probability of the null hypothesis (no difference) for  a composition (c). This is not a p-value.
  \item c_FDR - False-discovery rate of the null hypothesis (no difference) for  a composition (c).

  \item v_lower - (optional, present if variability is modelled dependent on covariates) lower (2.5%) quantile of the posterior distribution for a variability (v) parameter
  \item v_effect - (optional, present if variability is modelled dependent on covariates) mean of the posterior distribution for a variability (v) parameter
  \item v_upper - (optional, present if variability is modelled dependent on covariates) upper (97.5%) quantile of the posterior distribution for a variability (v) parameter
  \item v_pH0 - (optional, present if variability is modelled dependent on covariates) Probability of the null hypothesis (no difference) for a variability (v). This is not a p-value.
  \item v_FDR - (optional, present if variability is modelled dependent on covariates) False-discovery rate of the null hypothesis (no difference), for a variability (v).
}
}
\description{
The function for convoluted linear modelling takes as input a tidy table of feature count with three columns containing a sample ID, transcript ID and count, formula (continuous or discrete) and the covariate columns. The user can define a linear model with an input R formula, where the first covariate is the factor of interest.
}
\examples{

data("test_mixture")
data("no_hierarchy_reference")

 test_mixture |>
 convoluted_glm(
   ~ factor_of_interest,
   .sample = sample,
   .transcript = symbol,
   .abundance = count,
   reference = no_hierarchy_reference
  )

}
