% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{create_temp_result_directory}
\alias{create_temp_result_directory}
\title{Creates the directory needed for archive}
\usage{
create_temp_result_directory(name = "ARMET_results")
}
\arguments{
\item{name}{A char}
}
\description{
Creates the directory needed for archive
}
