% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_names_polar}
\alias{ct_names_polar}
\title{Abbreviations for magnolia plot}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 32 rows and 2 columns.
}
\usage{
ct_names_polar
}
\description{
Abbreviations for magnolia plot
}
\keyword{datasets}
