% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_NON_hierarchical.R,
%   R/functions_hierarchical.R
\name{estimate_convoluted_lm}
\alias{estimate_convoluted_lm}
\alias{estimate_convoluted_lm_1}
\alias{estimate_convoluted_lm_2}
\alias{estimate_convoluted_lm_3}
\alias{estimate_convoluted_lm_4}
\title{estimate_convoluted_lm}
\usage{
estimate_convoluted_lm(armet_obj, use_data = TRUE, use_cmdstanr = FALSE)

estimate_convoluted_lm_1(armet_obj)

estimate_convoluted_lm_2(armet_obj)

estimate_convoluted_lm_3(armet_obj)

estimate_convoluted_lm_4(armet_obj)
}
\arguments{
\item{armet_obj}{An ARMET object}
}
\description{
This function does inference for higher levels of the hierarchy

This function does inference for higher levels of the hierarchy

This function does inference for higher levels of the hierarchy

This function does inference for higher levels of the hierarchy

This function does inference for higher levels of the hierarchy
}
