% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{get_child_and_group_background}
\alias{get_child_and_group_background}
\title{Get the right level of the tree and label the leaf}
\usage{
get_child_and_group_background(node, ct)
}
\description{
Get the right level of the tree and label the leaf
}
