% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{get_distributions_from_tree}
\alias{get_distributions_from_tree}
\title{Get information from tree recursively for every node}
\usage{
get_distributions_from_tree(node)
}
\description{
Get information from tree recursively for every node
}
