% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{get_genes}
\alias{get_genes}
\title{Get genes from tree}
\usage{
get_genes(node, label = "markers", recursive = T, level = 0,
  start_level = 0, stop_level = 0)
}
\description{
Get genes from tree
}
