% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{get_leave_label}
\alias{get_leave_label}
\title{Get any info from the tree}
\usage{
get_leave_label(node, recursive = T, last_level = 0, label = "name")
}
\description{
Get any info from the tree
}
