% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{get_map_foreground_background}
\alias{get_map_foreground_background}
\title{Create a map where every leaf is mapped with the right ancestor}
\usage{
get_map_foreground_background(node, ct)
}
\description{
Create a map where every leaf is mapped with the right ancestor
}
