% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{get_mean_signature}
\alias{get_mean_signature}
\title{Averages the signatures of the same cell type}
\usage{
get_mean_signature(df, do_log = F, verbose = T)
}
\arguments{
\item{df}{A matrix}

\item{do_log}{A bool}

\item{verbose}{A bool}
}
\value{
A averaged matrix
}
\description{
Averages the signatures of the same cell type
}
