% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{get_node_label_level_specfic}
\alias{get_node_label_level_specfic}
\title{Get cell type names from a specific level of the tree}
\usage{
get_node_label_level_specfic(node, label = "name", recursive = T,
  level = 0, start_level = 0, stop_level = 0)
}
\description{
Get cell type names from a specific level of the tree
}
