% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tree.R
\name{get_node_label_recursive}
\alias{get_node_label_recursive}
\title{Get information from tree recursively for every node}
\usage{
get_node_label_recursive(node, last_level = 0, label = "name")
}
\description{
Get information from tree recursively for every node
}
