% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_overlap_descriptive_stats}
\alias{get_overlap_descriptive_stats}
\title{Print which reference genes are in the mix}
\usage{
get_overlap_descriptive_stats(mix_tbl, ref_tbl)
}
\description{
Print which reference genes are in the mix
}
