% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{get_proportions_array}
\alias{get_proportions_array}
\title{Get proportions from tree}
\usage{
get_proportions_array(node, sample, label = "absolute_proportion")
}
\description{
Get proportions from tree
}
