% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hierarchy_reference}
\alias{hierarchy_reference}
\title{Reference signatures}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1147994 rows and 5 columns.
}
\usage{
hierarchy_reference
}
\description{
Reference signatures
}
\keyword{datasets}
