% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{house_keeping_genes}
\alias{house_keeping_genes}
\title{House keeping genes list}
\format{
An object of class \code{character} of length 553.
}
\usage{
house_keeping_genes
}
\description{
House keeping genes list
}
\keyword{datasets}
