% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{level_to_plot_inferred_vs_observed}
\alias{level_to_plot_inferred_vs_observed}
\title{This is a generalisation of ifelse that acceots an object and return an objects}
\usage{
level_to_plot_inferred_vs_observed(result, level, cores = 20)
}
\description{
This is a generalisation of ifelse that acceots an object and return an objects
}
