% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{mds_plot}
\alias{mds_plot}
\title{Plots MDSplot from limma}
\usage{
mds_plot(mix, ref, cell_types)
}
\arguments{
\item{mix}{A matrix}

\item{ref}{A matrix}

\item{cell_types}{A char vector}
}
\description{
Plots MDSplot from limma
}
