% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n_markers}
\alias{n_markers}
\title{Number of markers to select from the reference}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 75 rows and 3 columns.
}
\usage{
n_markers
}
\description{
Number of markers to select from the reference
}
\keyword{datasets}
