% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{no_hierarchy_reference}
\alias{no_hierarchy_reference}
\title{Reference signatures}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 736364 rows and 4 columns.
}
\usage{
no_hierarchy_reference
}
\description{
Reference signatures
}
\keyword{datasets}
