% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{parse_result_2D}
\alias{parse_result_2D}
\title{Parse fit object in a 2D matrix}
\usage{
parse_result_2D(fit, param, column = 1)
}
\arguments{
\item{fit}{A stan fit object}

\item{param}{A char}

\item{column}{A real}
}
\description{
Parse fit object in a 2D matrix
}
