% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{parse_summary_vector_in_2D}
\alias{parse_summary_vector_in_2D}
\title{Parse a vector from stan fit object in a 2D matrix}
\usage{
parse_summary_vector_in_2D(f)
}
\arguments{
\item{f}{A array}
}
\value{
A data frame
}
\description{
Parse a vector from stan fit object in a 2D matrix
}
