% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_tc.R
\name{plot_counts_inferred_sum}
\alias{plot_counts_inferred_sum}
\title{plot_counts_inferred_sum}
\usage{
plot_counts_inferred_sum(fit_obj, samples = NULL)
}
\description{
Get data format for MPI deconvolution part
}
