% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities_plotting.R
\name{plot_densities}
\alias{plot_densities}
\title{Plots the densities distributions columnwise of a matrix}
\usage{
plot_densities(df, color = "0", fill = "0", alpha = 0.3, do_log = T)
}
\arguments{
\item{df}{A matrix}

\item{color}{A char}

\item{fill}{A array of numbers}

\item{alpha}{A number}

\item{do_log}{A bool}
}
\value{
A ggplot
}
\description{
Plots the densities distributions columnwise of a matrix
}
