% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities_plotting.R
\name{plot_densities_from_melted}
\alias{plot_densities_from_melted}
\title{Plots the densities from a melted data frame}
\usage{
plot_densities_from_melted(df_melted)
}
\arguments{
\item{df_melted}{A matrix}
}
\value{
A ggplot
}
\description{
Plots the densities from a melted data frame
}
