% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_tc.R
\name{plot_differences_in_lambda}
\alias{plot_differences_in_lambda}
\title{Plot differences between inferred and observed transcription abundances}
\usage{
plot_differences_in_lambda()
}
\description{
Plot differences between inferred and observed transcription abundances
}
