% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tree.R
\name{plot_makers}
\alias{plot_makers}
\title{Plot violin plot for markers in a node}
\usage{
plot_makers(tree, ct, df, do_replacement = F, markers = NULL)
}
\description{
Plot violin plot for markers in a node
}
