% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{plot_makers_in_tree}
\alias{plot_makers_in_tree}
\title{Plot violin plot for a node downstream}
\usage{
plot_makers_in_tree(tree, df, do_replacement = F)
}
\description{
Plot violin plot for a node downstream
}
