% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_polar}
\alias{plot_polar}
\title{plot_polar}
\usage{
plot_polar(
  .data,
  size_geom_text = 3.5,
  my_breaks = c(0, 0.01, 0.03, 0.05, 0.1, 0.3, 0.5, 0.7, 1),
  prop_filter = 0.005,
  barwidth = 0.5,
  barheight = 2,
  legend_justification = 0.67,
  fill_direction = 1
)
}
\arguments{
\item{.data}{ARMET-tc object}

\item{size_geom_text}{A double}

\item{my_breaks}{An integer}

\item{prop_filter}{A double}

\item{barwidth}{A double}

\item{barheight}{A double}

\item{legend_justification}{A double}

\item{fill_direction}{An integer}
}
\value{
a ggplot
}
\description{
Create polar plot of results
}
