% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{prepare_input}
\alias{prepare_input}
\title{Create object for full bayesian model, with the reference samples}
\usage{
prepare_input(ref, node)
}
\arguments{
\item{ref}{A matrix}

\item{cell_types}{A char rvector}

\item{markers}{A char vector}
}
\value{
A list including a two objects
}
\description{
Create object for full bayesian model, with the reference samples
}
