% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tree.R
\name{prop_array_to_tree}
\alias{prop_array_to_tree}
\title{Put proportions of cell types into the tree}
\usage{
prop_array_to_tree(tree, p_array)
}
\description{
Put proportions of cell types into the tree
}
