% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tree.R
\name{prop_table_to_leveled_prop_table}
\alias{prop_table_to_leveled_prop_table}
\title{Get prop table with proportions summarising at a selected level from another table}
\usage{
prop_table_to_leveled_prop_table(tree, df, level = 999)
}
\description{
Get prop table with proportions summarising at a selected level from another table
}
