% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{prop_treeList_to_leveled_prop_table}
\alias{prop_treeList_to_leveled_prop_table}
\title{Get prop table with proportions summarising at a selected level from trees}
\usage{
prop_treeList_to_leveled_prop_table(treeList, level = 999)
}
\description{
Get prop table with proportions summarising at a selected level from trees
}
