% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{quant_norm_to_array}
\alias{quant_norm_to_array}
\title{Normalize RNAseq to match array}
\usage{
quant_norm_to_array(obj, target)
}
\arguments{
\item{obj}{A matrix}

\item{target}{A matrix}
}
\value{
A list including the ref and mix normalized and a ggplot
}
\description{
Normalize RNAseq to match array
}
