% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{quant_norm_to_RNAseq}
\alias{quant_norm_to_RNAseq}
\title{Normalize array to match RNAseq}
\usage{
quant_norm_to_RNAseq(target, obj)
}
\arguments{
\item{target}{A matrix}

\item{obj}{A matrix}
}
\value{
A list including the ref and mix normalized and a ggplot
}
\description{
Normalize array to match RNAseq
}
