% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities_normalization.R
\name{quant_norm_to_target}
\alias{quant_norm_to_target}
\title{Normalize array to match RNAseq}
\usage{
quant_norm_to_target(obj, target)
}
\arguments{
\item{obj}{A matrix}

\item{target}{A matrix}
}
\value{
A list including the ref and mix normalized and a ggplot
}
\description{
Normalize array to match RNAseq
}
