% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_intercept_only.R
\name{ref_intercept_only}
\alias{ref_intercept_only}
\title{ref_intercept_only}
\usage{
ref_intercept_only(reference, level, cores = 8, approximate_posterior = F)
}
\arguments{
\item{mix}{A matrix}

\item{my_design}{A matrix}

\item{cov_to_test}{A character string}

\item{fully_bayesian}{A boolean}

\item{is_mix_microarray}{A boolean}

\item{verbose}{A boolean}

\item{save_report}{A boolean}

\item{custom_ref}{A matrix}

\item{multithread}{A boolean}

\item{do_debug}{A boolean}

\item{cell_type_root}{A character string}

\item{choose_internal_ref}{A design matrix}

\item{omit_regression}{A boolean}

\item{save_fit}{A boolean}

\item{seed}{An integer}
}
\value{
An ARMET object
}
\description{
This function calls the stan model.
}
