% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities_normalization.R
\name{rnaseq_norm.get_cpm}
\alias{rnaseq_norm.get_cpm}
\title{Eliminate lowly tanscribed genes for normalization}
\usage{
rnaseq_norm.get_cpm(df, cpm_threshold = 0.5)
}
\arguments{
\item{df}{A matrix}

\item{cpm_threshold}{A number}

\item{prop}{A number}
}
\value{
A tibble filtered
}
\description{
Eliminate lowly tanscribed genes for normalization
}
