% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities_normalization.R
\name{rnaseq_norm}
\alias{rnaseq_norm}
\title{Normalize a RNA seq data set using rnaseq_norm.calcNormFactor}
\usage{
rnaseq_norm(df, reference = NULL, cpm_threshold = 0.5, prop = 3/4,
  genes_to_keep = c())
}
\arguments{
\item{df}{A matrix}

\item{reference}{A reference matrix}

\item{cpm_threshold}{A number}

\item{prop}{A number}
}
\value{
A list including the filtered data frame and the normalization factors
}
\description{
Normalize a RNA seq data set using rnaseq_norm.calcNormFactor
}
