% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{run_coreAlg_though_tree_recursive}
\alias{run_coreAlg_though_tree_recursive}
\title{Run ARMET core algorithm recursively on the tree}
\usage{
run_coreAlg_though_tree_recursive(node, obj.in, bg_tree, log.ARMET)
}
\arguments{
\item{node}{A node from the tree}

\item{obj.in}{A object with the input}

\item{bg_tree}{The background tree that tracks the evolution of the algoritm}
}
\value{
A probability array
}
\description{
Run ARMET core algorithm recursively on the tree
}
