% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{run_test_though_tree}
\alias{run_test_though_tree}
\title{Run statstical test recursively in parallel}
\usage{
run_test_though_tree(node, my_design, cov_to_test)
}
\arguments{
\item{node}{A node from the tree}
}
\value{
A tree object
}
\description{
Run statstical test recursively in parallel
}
