% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{sanity_check_p}
\alias{sanity_check_p}
\title{Plots the chains for checking convergence}
\usage{
sanity_check_p(fit, do_show = F)
}
\arguments{
\item{fit}{A stan object}

\item{do_show}{A bool}
}
\value{
A ggplot
}
\description{
Plots the chains for checking convergence
}
