% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{setup_convolved_lm}
\alias{setup_convolved_lm}
\title{ARMET-tc main}
\usage{
setup_convolved_lm(
  .data,
  .formula = ~1,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  approximate_posterior = F,
  prior_survival_time = c(),
  transform_time_function = sqrt,
  reference = NULL
)
}
\arguments{
\item{.data}{A tibble}

\item{.formula}{A formula}

\item{.sample}{A column symbol}

\item{.transcript}{A column symbol}

\item{.abundance}{A column symbol}

\item{approximate_posterior}{A boolean for variational Bayes}

\item{prior_survival_time}{An array}

\item{transform_time_function}{transformation of the time covariate}

\item{reference}{A tibble}
}
\value{
An ARMET object
}
\description{
This function calls the stan model.
}
