% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities_hypothesisTest.R
\name{simplexRegression}
\alias{simplexRegression}
\title{Wrapper for the model and hipothesis testing of simplex regression}
\usage{
simplexRegression(proportions, my_design, cov_to_test = NULL, verbose = T)
}
\arguments{
\item{proportions}{simplex matrix (rows must sum to 1)}

\item{my_design}{design matrix}

\item{cov_to_test}{character variable indicating the column name of the design matrix to test}
}
\value{
a stan fit object
}
\description{
Wrapper for the model and hipothesis testing of simplex regression
}
\examples{
  simplexRegression(prop, my_design, "cov")
 
}
