% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{subsample_ref}
\alias{subsample_ref}
\title{Subsample reference to try homogeneity}
\usage{
subsample_ref(ref, cell_types, dim_mix = 10)
}
\arguments{
\item{ref}{A matrix}

\item{cell_types}{A char vector}

\item{dim_mix}{A real}
}
\value{
A numerical array of the selected samples
}
\description{
Subsample reference to try homogeneity
}
