% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_tc.R
\name{test_differential_composition}
\alias{test_differential_composition}
\title{test_differential_composition}
\usage{
test_differential_composition(
  .data,
  credible_interval = 0.9,
  cluster_CI = 0.55,
  relative = TRUE
)
}
\arguments{
\item{.data}{A tibble}

\item{credible_interval}{A double}

\item{cluster_CI}{A double}
}
\description{
This function performs statistical testing on the fit object
}
