% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tree.R
\name{test_genes_in_node}
\alias{test_genes_in_node}
\title{Test differentially transcribed gene for a node}
\usage{
test_genes_in_node(tree, df, ct, n = 20, do_replacement = F,
  ct_to_test_against = NULL, expression_threshold = 100, include_bg = F,
  quantile_norm = T, na_rm = F, do_higher = T, do_calc_sd = T)
}
\description{
Test differentially transcribed gene for a node
}
