% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tree.R
\name{test_genes_in_tree}
\alias{test_genes_in_tree}
\title{Get top differentially transcribed in all tree}
\usage{
test_genes_in_tree(tree, df, n = 20, do_replacement = F,
  expression_threshold = 100, na_rm = F)
}
\description{
Get top differentially transcribed in all tree
}
