% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{test_hypothesis_convoluted_lm}
\alias{test_hypothesis_convoluted_lm}
\title{test_hypothesis_convoluted_lm}
\usage{
test_hypothesis_convoluted_lm(.data, CI = 0.9, relative = TRUE)
}
\arguments{
\item{.data}{A tibble}

\item{CI}{A double}
}
\description{
This function performs statistical testing on the fit object
}
