% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_tc.R
\name{ToDataFrameTypeColFull}
\alias{ToDataFrameTypeColFull}
\title{ToDataFrameTypeColFull}
\usage{
ToDataFrameTypeColFull(tree, ...)
}
\description{
Extension of data.tree package. It converts the tree into data frame
}
