% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{tree_to_leveled_prop}
\alias{tree_to_leveled_prop}
\title{Get prop table with proportions summarising at a selected level from a tree}
\usage{
tree_to_leveled_prop(tree, level = 999)
}
\description{
Get prop table with proportions summarising at a selected level from a tree
}
