% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vb_iterative}
\alias{vb_iterative}
\title{vb_iterative}
\usage{
vb_iterative(
  model,
  output_samples = 2000,
  iter = 10000,
  tol_rel_obj = 0.01,
  additional_parameters_to_save = list(),
  init,
  data,
  ...
)
}
\arguments{
\item{model}{A Stan model}

\item{output_samples}{An integer of how many samples from posteriors}

\item{iter}{An integer of how many max iterations}

\item{tol_rel_obj}{A real}

\item{additional_parameters_to_save}{A character vector}

\item{init}{A list}

\item{data}{A list}

\item{...}{List of paramaters for vb function of Stan}
}
\value{
A Stan fit object
}
\description{
Runs iteratively variational bayes until it suceeds
}
