% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities_normalization.R
\name{wrapper_normalize_mix_ref}
\alias{wrapper_normalize_mix_ref}
\title{Wrapper function for all normalization based on the current needs}
\usage{
wrapper_normalize_mix_ref(mix, ref, is_mix_microarray)
}
\arguments{
\item{mix}{A matrix}

\item{ref}{A matrix}

\item{is_mix_microarray}{A cool}
}
\value{
Same list of rnaseq_norm_ref_mix, quant_norm_to_target
}
\description{
Wrapper function for all normalization based on the current needs
}
