% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{add_absolute_proportions_to_tree}
\alias{add_absolute_proportions_to_tree}
\title{Put absolute proportions on tree}
\usage{
add_absolute_proportions_to_tree(node, p_ancestor = NULL)
}
\description{
Put absolute proportions on tree
}
