\name{ARMET-package}
\alias{ARMET-package}
\alias{ARMET}
\docType{package}
\title{
\packageTitle{ARMET}
}
\description{
\packageDescription{ARMET}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{ARMET}
\packageIndices{ARMET}
bla
}
\author{
\packageAuthor{ARMET}
Maintainer: \packageMaintainer{ARMET}
}
\references{
bla
}
bla
\keyword{ package }
\seealso{
bla
}
\examples{
bla
}
