% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET-package.R
\docType{package}
\name{ARMET-package}
\alias{ARMET-package}
\alias{ARMET}
\title{The 'ARMET' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
