% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_tc.R
\name{ARMET_tc}
\alias{ARMET_tc}
\title{ARMET-tc main}
\usage{
ARMET_tc(mix, my_design = NULL, cov_to_test = NULL, fully_bayesian = F,
  is_mix_microarray = F, observed_prop = NULL,
  ct_to_omit = c("t_CD4_naive", "adipocyte"), verbose = F,
  sigma_hyper_sd = 0.1, phi_hyper_sd = 2, alpha_hyper_value = 2,
  save_report = F, custom_ref = NULL, multithread = T, do_debug = F,
  cell_type_root = "TME", choose_internal_ref = NULL, omit_regression = F,
  save_fit = F, seed = NULL)
}
\arguments{
\item{mix}{A matrix}

\item{my_design}{A matrix}

\item{cov_to_test}{A character string}

\item{fully_bayesian}{A boolean}

\item{is_mix_microarray}{A boolean}

\item{observed_prop}{A matrix}

\item{ct_to_omit}{A character string}

\item{verbose}{A boolean}

\item{sigma_hyper_sd}{A real}

\item{phi_hyper_sd}{A real}

\item{alpha_hyper_value}{A real}

\item{save_report}{A boolean}

\item{custom_ref}{A matrix}

\item{multithread}{A boolean}

\item{do_debug}{A boolean}

\item{cell_type_root}{A character string}

\item{choose_internal_ref}{A design matrix}

\item{omit_regression}{A boolean}

\item{save_fit}{A boolean}

\item{seed}{An integer}
}
\value{
An ARMET object
}
\description{
This function calls the stan model.
}
