% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_tc.R
\name{ARMET_tc}
\alias{ARMET_tc}
\title{ARMET-tc main}
\usage{
ARMET_tc(mix, reference, ct_to_omit = c("t_CD4_naive", "adipocyte"),
  verbose = F, omit_regression = F, save_fit = F, seed = NULL,
  cores = 14, iterations = 300, sampling_iterations = 100,
  levels = 1:4, full_bayes = T, n_markers)
}
\arguments{
\item{mix}{A matrix}

\item{ct_to_omit}{A character string}

\item{verbose}{A boolean}

\item{omit_regression}{A boolean}

\item{save_fit}{A boolean}

\item{seed}{An integer}

\item{my_design}{A matrix}

\item{cov_to_test}{A character string}

\item{fully_bayesian}{A boolean}

\item{is_mix_microarray}{A boolean}

\item{save_report}{A boolean}

\item{custom_ref}{A matrix}

\item{multithread}{A boolean}

\item{do_debug}{A boolean}

\item{cell_type_root}{A character string}

\item{choose_internal_ref}{A design matrix}
}
\value{
An ARMET object
}
\description{
This function calls the stan model.
}
