% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_tc.R
\name{ARMET_tc}
\alias{ARMET_tc}
\title{ARMET-tc main}
\usage{
ARMET_tc(
  .data,
  .formula = ~1,
  .sample = NULL,
  .transcript = NULL,
  .abundance = NULL,
  reference = NULL,
  approximate_posterior = F,
  verbose = F,
  save_fit = F,
  seed = NULL,
  cores = 14,
  iterations = 250,
  sampling_iterations = 100,
  levels = 3,
  n_markers = my_n_markers,
  do_regression = F
)
}
\arguments{
\item{verbose}{A boolean}

\item{save_fit}{A boolean}

\item{seed}{An integer}

\item{mix}{A matrix}

\item{my_design}{A matrix}

\item{cov_to_test}{A character string}

\item{fully_bayesian}{A boolean}

\item{is_mix_microarray}{A boolean}

\item{save_report}{A boolean}

\item{custom_ref}{A matrix}

\item{multithread}{A boolean}

\item{do_debug}{A boolean}

\item{cell_type_root}{A character string}

\item{choose_internal_ref}{A design matrix}
}
\value{
An ARMET object
}
\description{
This function calls the stan model.
}
