% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_tc.R
\name{parse_summary_check_divergence}
\alias{parse_summary_check_divergence}
\title{parse_summary_check_divergence}
\usage{
parse_summary_check_divergence(fit)
}
\description{
Parse the stan fit object and check for divergencies
}
