% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_tc.R
\name{plot_markers}
\alias{plot_markers}
\title{This is a generalisation of ifelse that acceots an object and return an objects}
\usage{
plot_markers(result, level, S = NULL, cores = 20)
}
\description{
This is a generalisation of ifelse that acceots an object and return an objects
}
