% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities.R
\name{rnaseq_norm.calcNormFactor}
\alias{rnaseq_norm.calcNormFactor}
\title{Calculate the norm factor with calcNormFactor from limma}
\usage{
rnaseq_norm.calcNormFactor(df, reference = NULL, cpm_theshold = 0.5,
  prop = 3/4)
}
\arguments{
\item{df}{A matrix}

\item{reference}{A reference matrix}

\item{cpm_theshold}{A number}

\item{prop}{A number}
}
\value{
A list including the filtered data frame and the normalization factors
}
\description{
Calculate the norm factor with calcNormFactor from limma
}
