% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_utilities_normalization.R
\name{rnaseq_norm_ref_mix}
\alias{rnaseq_norm_ref_mix}
\title{Normalize ref to match mix using TMM}
\usage{
rnaseq_norm_ref_mix(obj, target)
}
\arguments{
\item{ref}{A matrix}

\item{mix}{A matrix}
}
\value{
A list including the ref and mix normalized and a ggplot
}
\description{
Normalize ref to match mix using TMM
}
