% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMET_parseTree.R
\name{run_coreAlg_though_tree}
\alias{run_coreAlg_though_tree}
\title{Run ARMET core algorithm recursively on the tree}
\usage{
run_coreAlg_though_tree(node, obj.in)
}
\arguments{
\item{node}{A node from the tree}

\item{obj.in}{A object with the input}
}
\value{
A probability array
}
\description{
Run ARMET core algorithm recursively on the tree
}
